pyproject.toml
[build-system]
requires = ["poetry>=1.1"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "upperroom"
description = "A website for Churches."
authors = [
    "Craig Holyoak <craig@helmsdeep.org>",
]
license = "GPL-2.0-only"
readme = "README.md"
repository = "https://github.com/thepointchurch/upperroom"
homepage = "https://github.com/thepointchurch/upperroom"
version = "2.4.6"

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

exclude = [
    "**/tests.py",
    "**/tests",
]

[tool.poetry.dependencies]
python = "^3.7"

argon2-cffi = "^20.1"
bcrypt = "^3.1"
django = "^3.1"
django-csp = "^3.7"
django-environ = "^0.4"
django-markwhat = "^1.6"
django-robots = "^4.0"
django-sendmail-backend = "^0.1"
mutagen = "^1.45"
weasyprint = "^51"
yafg = "^0.3"
importlib-metadata = { version = "^1.7", python = "<3.8" }

gunicorn = "^20.0"
setproctitle = "^1.1"

cryptography = { version = "^3.3", optional = true }
django-amazon-ses = { version = "^4.0", optional = true }
django-storages = { version = "^1.9", optional = true }
google-api-python-client = { version = "^1.10", optional = true }
google-auth-oauthlib = { version = "^0.4", optional = true }
psycopg2 = { version = "^2.8", optional = true }
redis = { version = "^3.5", optional = true }
django-redis = { version = "^4.12", optional = true }
django-cacheops = { version = "^5.0", optional = true }

[tool.poetry.extras]
aws = [
    "cryptography",
    "django-amazon-ses",
    "django-storages",
]
cache = [
    "django-cacheops",
    "django-redis",
]
google = [
    "google-api-python-client",
    "google-auth-oauthlib",
]
pgsql = [
    "psycopg2",
]

[tool.poetry.dev-dependencies]
coverage = "5.3"
django-debug-toolbar = "^2.2"
gitpython = "^3.1"
pre-commit = "^2.6"
pylint = "^2.5"
pylint-django = "^2.3"
pytest = "^6.0"
pytest-django = "^3.9"
semantic_version = "^2.8"
tomlkit = "^<1.0"

[tool.poetry.scripts]
upperroom = 'upperroom.commands:main'
upperroom-sendrosteremails = 'upperroom.commands:sendrosteremails'
upperroom-restore-fixtures = 'upperroom.commands:restore_fixtures'


[tool.pylint.master]
jobs = 0
load-plugins = [
    "pylint_django",
]
django-settings-module = "upperroom.settings"

[tool.pylint.format]
max-line-length = 119

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "bad-continuation",
    "bad-whitespace",
    "duplicate-code",
]

[tool.pylint.variables]
ignored-argument-names = "kwargs|args"


[tool.isort]
combine_as_imports = true
include_trailing_comma = true
known_first_party = "upperroom"
line_length = 119
multi_line_output = 3
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true


[tool.black]
line-length = 119
include = '\.pyi?$'


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
DJANGO_SETTINGS_MODULE = "upperroom.settings"
python_files = "tests.py tests/*.py"
filterwarnings = "ignore::django.utils.deprecation.RemovedInDjango40Warning"
